+++ src/main/java/com/dam/kanpeki/service/impl/UserServiceImpl.java
@@ -4,24 +4,22 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
-import com.dam.kanpeki.exception.DataNotFoundException;
-import com.dam.kanpeki.model.Question;
-import com.dam.kanpeki.model.dto.RequestUserDTO;
-import com.dam.kanpeki.model.dto.ResponseUserDTO;
-import com.dam.kanpeki.model.dto.mapper.UserDTOMapperStruct;
-import com.dam.kanpeki.service.FileSystemStorageServiceI;
-import com.dam.kanpeki.service.ResultServiceI;
-import com.dam.kanpeki.utils.FileUtils;
-import com.dam.kanpeki.utils.KanpekiConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
-
-import com.dam.kanpeki.model.User;
-import com.dam.kanpeki.repository.UserRepository;
-import com.dam.kanpeki.service.UserServiceI;
 import org.springframework.web.multipart.MultipartFile;
+
+import com.dam.kanpeki.exception.DataNotFoundException;
+import com.dam.kanpeki.model.User;
+import com.dam.kanpeki.model.dto.RequestUserDTO;
+import com.dam.kanpeki.model.dto.ResponseUserDTO;
+import com.dam.kanpeki.model.dto.mapper.UserDTOMapperStruct;
+import com.dam.kanpeki.repository.UserRepository;
+import com.dam.kanpeki.service.FileSystemStorageServiceI;
+import com.dam.kanpeki.service.ResultServiceI;
+import com.dam.kanpeki.service.UserServiceI;
+import com.dam.kanpeki.utils.KanpekiConstants;
 
 @Service
 public class UserServiceImpl implements UserServiceI {
@@ -72,9 +70,12 @@ public class UserServiceImpl implements UserServiceI {
 		u.setNickname(uField);
 
 		ExampleMatcher customExMatcher = ExampleMatcher.matchingAny()
-				.withMatcher(KanpekiConstants.USER_EMAIL_NAME, ExampleMatcher.GenericPropertyMatchers.contains().ignoreCase())
-				.withMatcher(KanpekiConstants.USER_FULLNAME_NAME, ExampleMatcher.GenericPropertyMatchers.contains().ignoreCase())
-				.withMatcher(KanpekiConstants.USER_NICKNAME_NAME, ExampleMatcher.GenericPropertyMatchers.contains().ignoreCase());
+				.withMatcher(KanpekiConstants.USER_EMAIL_NAME,
+						ExampleMatcher.GenericPropertyMatchers.contains().ignoreCase())
+				.withMatcher(KanpekiConstants.USER_FULLNAME_NAME,
+						ExampleMatcher.GenericPropertyMatchers.contains().ignoreCase())
+				.withMatcher(KanpekiConstants.USER_NICKNAME_NAME,
+						ExampleMatcher.GenericPropertyMatchers.contains().ignoreCase());
 
 		Example<User> userExample = Example.of(u, customExMatcher);
 
@@ -84,39 +85,45 @@ public class UserServiceImpl implements UserServiceI {
 	@Override
 	public Optional<ResponseUserDTO> findById(Long id) {
 		Optional<User> opUser = uRepo.findById(id);
-		if(!opUser.isPresent()){
+
+		if (!opUser.isPresent()) {
 			throw new DataNotFoundException(KanpekiConstants.EMPTY_STRING);
 		} else {
-			return Optional.of(mapper.toUserDTO(uRepo.findById(id).get()));
+			return Optional.of(mapper.toUserDTO(opUser.get()));
 		}
 	}
 
 	@Override
 	public ResponseUserDTO addUser(RequestUserDTO u, MultipartFile file) {
 		User uTemp = mapper.requestUserDTOtoUser(u);
-		uTemp.setUrlImage(FileUtils.saveFileRequest(file));
+		uTemp.setUrlImage(storeService.saveFileRequest(file));
 		return mapper.toUserDTO(uRepo.save(uTemp));
 	}
 
 	@Override
 	public void removeUserById(Long id) {
 		Optional<User> u = uRepo.findById(id);
-		// Eliminamos la imagen del almacenamiento
-		storeService.delete(u.get().getUrlImage());
 
-		// Eliminamos sus resultados si los tuviera
-		if (!u.get().getResults().isEmpty()) {
-			rService.deleteResultsByUserId(id);
+		if (!u.isPresent()) {
+			throw new DataNotFoundException(KanpekiConstants.EMPTY_STRING);
+		} else {
+			// Eliminamos la imagen del almacenamiento
+			storeService.delete(u.get().getUrlImage());
+
+			// Eliminamos sus resultados si los tuviera
+			if (!u.get().getResults().isEmpty()) {
+				rService.deleteResultsByUserId(id);
+			}
+
+			uRepo.deleteById(id);
 		}
-
-		uRepo.deleteById(id);
 	}
 
 	@Override
 	public ResponseUserDTO updateUser(RequestUserDTO u, MultipartFile file, Long id) {
 
 		User mappedU = mapper.requestUserDTOtoUser(u);
-		mappedU.setUrlImage(FileUtils.saveFileRequest(file));
+		mappedU.setUrlImage(storeService.saveFileRequest(file));
 
 		User mappedUUpdated = uRepo.findById(id).map(newU -> {
 			newU.setEmail(mappedU.getEmail());
